Update, from the meeting.
